malloc and free

+++ 0x0B-malloc_free/1-strdup.c
@@ -7,9 +7,9 @@
  * Return: string length
  */
 
-int _strlen(char *s)
+size_t _strlen(char *s)
 {
-	int len = 0;
+	size_t len = 0;
 
 	while (s[len] != '\0')
 		len++;
